[CHORE]: Move symbols out of /ts-interfaces and into -private/utils
