+++ packages/store/addon/-private/system/store/record-data-store-wrapper.ts
@@ -2,7 +2,7 @@ import { CUSTOM_MODEL_CLASS } from '@ember-data/canary-features';
 
 import { identifierCacheFor } from '../../identifiers/cache';
 import { RecordDataStoreWrapper as IRecordDataStoreWrapper } from '../../ts-interfaces/record-data-store-wrapper';
-import { BRAND_SYMBOL } from '../../ts-interfaces/utils/brand';
+import { BRAND_SYMBOL } from '../../utils/brand';
 import constructResource from '../../utils/construct-resource';
 import { upgradeForInternal } from '../ts-upgrade-map';
 import { internalModelFactoryFor } from './internal-model-factory';
